Fix folder structure
