+++ Remotion/Core/Extensions.UnitTests/Logging/LoggerExtensionsObsoleteTests/LogTest.cs
@@ -18,9 +18,7 @@ using System;
 using log4net.Core;
 using NUnit.Framework;
 using Remotion.Logging;
-
-using MicrosoftLogLevel = Microsoft.Extensions.Logging.LogLevel;
-
+using MicrosoftLoglevel = Microsoft.Extensions.Logging.LogLevel;
 
 namespace Remotion.Extensions.UnitTests.Logging.LoggerExtensionsObsoleteTests
 {
@@ -28,40 +26,148 @@ namespace Remotion.Extensions.UnitTests.Logging.LoggerExtensionsObsoleteTests
   public class LogTest : BaseTest
   {
     [Test]
-    public void Test_FormatWithEnumAndException ()
+    public void Test_WithMessageAndEventID ()
     {
-      Exception exception = new Exception();
       SetLoggingThreshold(Level.Info);
 
-      Log.LogFormat(MicrosoftLogLevel.Information, LogMessages.TheMessage, exception, "First", "Second");
+      Log.Log(MicrosoftLoglevel.Information, 1, (object)"The message.");
 
       LoggingEvent[] events = GetLoggingEvents();
       Assert.That(events.Length, Is.EqualTo(1));
       LoggingEvent loggingEvent = events[0];
       Assert.That(loggingEvent.Level, Is.EqualTo(Level.Info));
-      Assert.That(loggingEvent.MessageObject.ToString(), Is.EqualTo($"The message with First and Second."));
-      // Assert.That(loggingEvent.Properties["EventID"], Is.EqualTo((int)LogMessages.TheMessage));
+      Assert.That(loggingEvent.MessageObject, Is.EqualTo("The message."));
+      Assert.That(loggingEvent.Properties["EventID"], Is.EqualTo(1));
+      Assert.That(loggingEvent.ExceptionObject, Is.Null);
+      Assert.That(loggingEvent.Repository, Is.SameAs(Logger.Repository));
+      Assert.That(loggingEvent.LoggerName, Is.EqualTo(Logger.Name));
+    }
+
+    [Test]
+    public void Test_WithMessageAndException ()
+    {
+      Exception exception = new Exception();
+      SetLoggingThreshold(Level.Info);
+
+      Log.Log(MicrosoftLoglevel.Information, (object)"The message.", exception);
+
+      LoggingEvent[] events = GetLoggingEvents();
+      Assert.That(events.Length, Is.EqualTo(1));
+      LoggingEvent loggingEvent = events[0];
+      Assert.That(loggingEvent.Level, Is.EqualTo(Level.Info));
+      Assert.That(loggingEvent.MessageObject, Is.EqualTo("The message."));
       Assert.That(loggingEvent.ExceptionObject, Is.SameAs(exception));
       Assert.That(loggingEvent.Repository, Is.SameAs(Logger.Repository));
       Assert.That(loggingEvent.LoggerName, Is.EqualTo(Logger.Name));
     }
 
     [Test]
-    public void Test_FormatWithEnum ()
+    public void Test_WithMessage ()
     {
       SetLoggingThreshold(Level.Info);
 
-      Log.LogFormat(MicrosoftLogLevel.Information, LogMessages.TheMessage, null, "First", "Second");
+      Log.Log(MicrosoftLoglevel.Information, (object)"The message.");
 
       LoggingEvent[] events = GetLoggingEvents();
       Assert.That(events.Length, Is.EqualTo(1));
       LoggingEvent loggingEvent = events[0];
       Assert.That(loggingEvent.Level, Is.EqualTo(Level.Info));
-      Assert.That(loggingEvent.MessageObject.ToString(), Is.EqualTo("The message with First and Second."));
-      // Assert.That(loggingEvent.Properties["EventID"], Is.EqualTo((int)LogMessages.TheMessage));
+      Assert.That(loggingEvent.MessageObject, Is.EqualTo("The message."));
       Assert.That(loggingEvent.ExceptionObject, Is.Null);
       Assert.That(loggingEvent.Repository, Is.SameAs(Logger.Repository));
       Assert.That(loggingEvent.LoggerName, Is.EqualTo(Logger.Name));
+    }
+
+    [Test]
+    public void Test_WithLogLevelNone ()
+    {
+      Logger.Repository.Threshold = Level.Off;
+
+      Log.Log(MicrosoftLoglevel.Information, 1, (object)"The message.");
+
+      Assert.That(GetLoggingEvents(), Is.Empty);
+    }
+
+    [Test]
+    public void Test_FormatWithMessageAndEventIDAndException ()
+    {
+      Exception exception = new Exception();
+      SetLoggingThreshold(Level.Info);
+
+      Log.InfoFormat(1, exception, "{0} {1}", "The", "message.");
+
+      LoggingEvent[] events = GetLoggingEvents();
+      Assert.That(events.Length, Is.EqualTo(1));
+      LoggingEvent loggingEvent = events[0];
+      Assert.That(loggingEvent.Level, Is.EqualTo(Level.Info));
+      Assert.That(loggingEvent.MessageObject.ToString(), Is.EqualTo("The message."));
+      Assert.That(loggingEvent.Properties["EventID"], Is.EqualTo(1));
+      Assert.That(loggingEvent.ExceptionObject, Is.SameAs(exception));
+      Assert.That(loggingEvent.Repository, Is.SameAs(Logger.Repository));
+      Assert.That(loggingEvent.LoggerName, Is.EqualTo(Logger.Name));
+    }
+
+    [Test]
+    public void Test_FormatWithMessageAndException ()
+    {
+      Exception exception = new Exception();
+      SetLoggingThreshold(Level.Info);
+
+      Log.LogFormat(MicrosoftLoglevel.Information, exception, "{0} {1}", "The", "message.");
+
+      LoggingEvent[] events = GetLoggingEvents();
+      Assert.That(events.Length, Is.EqualTo(1));
+      LoggingEvent loggingEvent = events[0];
+      Assert.That(loggingEvent.Level, Is.EqualTo(Level.Info));
+      Assert.That(loggingEvent.MessageObject.ToString(), Is.EqualTo("The message."));
+      Assert.That(loggingEvent.ExceptionObject, Is.SameAs(exception));
+      Assert.That(loggingEvent.Repository, Is.SameAs(Logger.Repository));
+      Assert.That(loggingEvent.LoggerName, Is.EqualTo(Logger.Name));
+    }
+
+    [Test]
+    public void Test_FormatWithMessageAndEventID ()
+    {
+      SetLoggingThreshold(Level.Info);
+
+      Log.LogFormat(MicrosoftLoglevel.Information, 1, "{0} {1}", "The", "message.");
+
+      LoggingEvent[] events = GetLoggingEvents();
+      Assert.That(events.Length, Is.EqualTo(1));
+      LoggingEvent loggingEvent = events[0];
+      Assert.That(loggingEvent.Level, Is.EqualTo(Level.Info));
+      Assert.That(loggingEvent.MessageObject.ToString(), Is.EqualTo("The message."));
+      Assert.That(loggingEvent.Properties["EventID"], Is.EqualTo(1));
+      Assert.That(loggingEvent.ExceptionObject, Is.Null);
+      Assert.That(loggingEvent.Repository, Is.SameAs(Logger.Repository));
+      Assert.That(loggingEvent.LoggerName, Is.EqualTo(Logger.Name));
+    }
+
+    [Test]
+    public void Test_FormatWithMessage ()
+    {
+      SetLoggingThreshold(Level.Info);
+
+      Log.InfoFormat("{0} {1}", "The", "message.");
+
+      LoggingEvent[] events = GetLoggingEvents();
+      Assert.That(events.Length, Is.EqualTo(1));
+      LoggingEvent loggingEvent = events[0];
+      Assert.That(loggingEvent.Level, Is.EqualTo(Level.Info));
+      Assert.That(loggingEvent.MessageObject.ToString(), Is.EqualTo("The message."));
+      Assert.That(loggingEvent.ExceptionObject, Is.Null);
+      Assert.That(loggingEvent.Repository, Is.SameAs(Logger.Repository));
+      Assert.That(loggingEvent.LoggerName, Is.EqualTo(Logger.Name));
+    }
+
+    [Test]
+    public void Test_FormatWithLogLevelNone ()
+    {
+      Logger.Repository.Threshold = Level.Off;
+
+      Log.LogFormat(MicrosoftLoglevel.Information, 1, "{0} {1}", "The", "message.");
+
+      Assert.That(GetLoggingEvents(), Is.Empty);
     }
   }
 }
